Don't define nginx_int_t et al as uintptr_t, its... wrong.  It's just wrong.

+++ src/core/ngx_config.h
@@ -75,9 +75,9 @@
 
 #endif
 
-typedef intptr_t        ngx_int_t;
-typedef uintptr_t       ngx_uint_t;
-typedef intptr_t        ngx_flag_t;
+typedef int64_t		ngx_int_t;
+typedef uint64_t	ngx_uint_t;
+typedef int64_t		ngx_flag_t;
 
 
 #define NGX_INT32_LEN   (sizeof("-2147483648") - 1)
